requirements.txt
requests
beautifulsoup4
feedparser
python-dotenv
tqdm
schedule
FastAPI
uvicorn
markdown
sqlalchemy
bcrypt
python-jose[cryptography]
python-multipart
jinja2
aiosmtplib
apscheduler
passlib[bcrypt]
fastapi-users[sqlalchemy]
httpx